requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataframe_image==0.2.7
DataRecorder==3.6.2
defusedxml==0.7.1
DownloadKit==2.0.7
et_xmlfile==2.0.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lxml==5.3.1
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
mistune==3.1.2
more-itertools==10.6.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
pillow==11.1.0
platformdirs==4.3.6
playwright==1.50.0
propcache==0.3.0
psutil==7.0.0
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1

pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rpds-py==0.23.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.39
tabulate==0.9.0
tinycss2==1.4.0
tldextract==5.1.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
yarl==1.18.3
